requirements.txt
torch
transformers
datasets
pycocotools
pillow
numpy
tqdm
accelerate
